requirements.txt
bitarray
bottleneck
brotlicffi
brotlipy
chardet
conllu
datasets
emoji
envisage
fastbpe
hydra-optuna-sweeper
jupyter
mkl-fft
mkl-random
mkl-service
mlflow
more-itertools
munkres
nlpaug
nltk
numexpr
olefile
openpyxl
pip-chill
pipreqs
py7zr
pyopenssl
pysocks
pytest
sacrebleu
sacremoses
seaborn
sentencepiece
setuptools-scm
sklearn
subword-nmt
torchaudio
torchvision
transformers
vtk
